Updates of (mainly) running changes

+++ @lvtime/uint8.m
@@ -7,4 +7,4 @@ x1=rem(x3,256);
 x3=floor(x3/256);
 x2=rem(x3,256);
 x3=floor(x3/256);
-x=reshape([x3;x2;x1;x0],1,16);
+x=uint8(reshape([x3;x2;x1;x0],1,16));
